Rebase and improve test coverage of this branch

+++ dlme_airflow/utils/princeton.py
@@ -2,17 +2,20 @@
 import os
 import pandas as pd
 
+from dlme_airflow.utils.catalog import get_working_csv
+
 
 def remove_ymdi(**kwargs):
     coll = kwargs["collection"]
-    root_dir = os.path.dirname(os.path.abspath("metadata"))
     data_path = coll.data_path()
-    working_csv = os.path.join(root_dir, "working", data_path, "data.csv")
-    df = pd.read_csv(working_csv)
+    working_csv = get_working_csv(data_path)
 
-    filter_df(df)
+    if os.path.isfile(working_csv):
+        df = pd.read_csv(working_csv)
+        filter_df(df)
+        df.to_csv(working_csv)
 
-    df.to_csv(working_csv)
+    return working_csv
 
 
 def filter_df(df):
